chore: refactor observations domain
