+++ web/src/features/public-api/types/observations.ts
@@ -5,7 +5,10 @@ import {
   publicApiPaginationZod,
 } from "@langfuse/shared";
 
-import { stringDateTime } from "@langfuse/shared/src/server";
+import {
+  reduceUsageOrCostDetails,
+  stringDateTime,
+} from "@langfuse/shared/src/server";
 import type Decimal from "decimal.js";
 import { z } from "zod";
 
@@ -81,29 +84,6 @@ export const APIObservation = z
 /**
  * Transforms
  */
-export const reduceUsageOrCostDetails = (
-  details: Record<string, number> | null | undefined,
-): {
-  input: number | null;
-  output: number | null;
-  total: number | null;
-} => {
-  return {
-    input: Object.entries(details ?? {})
-      .filter(([usageType]) => usageType.startsWith("input"))
-      .reduce(
-        (acc, [_, value]) => (acc ?? 0) + Number(value),
-        null as number | null, // default to null if no input usage is found
-      ),
-    output: Object.entries(details ?? {})
-      .filter(([usageType]) => usageType.startsWith("output"))
-      .reduce(
-        (acc, [_, value]) => (acc ?? 0) + Number(value),
-        null as number | null, // default to null if no output usage is found
-      ),
-    total: Number(details?.total ?? 0),
-  };
-};
 
 /**
  *
